Remove cli functionality from smileycoind
As it says on the tin. It was deprecated in bitcoind version 0.9, and
at some point it should be removed.

Removes the dependency of smileycoind on libbitcoin-cli.a. Move
some functions that used to be shared but are now only used in
bitcoin-cli.cpp to that file.

After this change, an error is printed (and exit code 1 is returned)
when the user tries to send RPC commands using smileycoind.